app/book_scraper/requirements.txt
python-dotenv
scrapy
psycopg2